requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
cycler==0.11.0
fonttools==4.37.1
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-spreadsheet==0.0.6
googleapis-common-protos==1.56.4
gspread==5.5.0
gspread-formatting==1.1.1
httplib2==0.20.4
idna==3.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tabulate==0.8.10
tornado==6.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
